[DON-79] Removed jest and all related items from the `./server/` folder.
Deleted old `/api/1/` *.test.ts files since we are migrating to
Playwright and the current server jest tests are very difficult to mock
(JWT authorization and Auth0) without hammering the API. Playwright
tests will need to be written for the new `/api/2/` endpoints and since
we are migrating to Prisma, no further development should be happening
on the `/api/1/` endpoints, so I feel it is safe for the time being to
go without tests for the server.

The next steps will be redoing the current OpenAPI setup for the
`./server/src/server.ts` file and allow creating and using an API key
that a user can request via username/password or by sending a
pregenerated API key in the header.

Currently, this setup will (hopefully) prevent any hammering of the
Auth0 api so we don't go over and don't have to pay for any extra API
calls/month.