[DON-151/auth0-server-jest-fix]
Jest lives again. A refactored auth0 process for the server has been created and now stores the oauth token for tests in the database. This requires users to get a second ID and Secret, but there is a spot in the .env.dist file for them, so users will need to either add the missing entries in their current .env or copy the .env.dist to .env and add any missing information then.

README.md has been updated to reflect the new changes and explains that the user needs to nuke their current database in order for this to work, as it adds a missing table for token storage.

WonderTix.code-workspace has removed the server from the ignore list for jest tests so they can run automatically in VSCode again when opening this file.