[DON-151/auth0-server-jest-fix]
Made some changes to the way the setup is done. There is an .env file
now in the ./server folder that needs to be edited (instructions in the
README.md) in order for the tests to run properly on linux/macOS systems.
This still needs tested on Windows a bit more to ensure it has no
regression errors, as it worked on Windows prior to this commit.